pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bunny-app"
version = "0.1.0"
description = "Bunny App service scaffold built from the Bunny2 realtime translation MVP."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "ctranslate2>=4.6.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "pywebpush>=2.0.3",
  "pydantic-settings>=2.8.0",
  "python-multipart>=0.0.20",
  "sentencepiece>=0.2.0",
  "transformers>=4.46.0",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.12.0",
]
postgres = [
  "psycopg[binary]>=3.2.0",
]

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["static/*"]

[tool.ruff]
line-length = 100
target-version = "py311"
